refactor(scan): split manager.go along natural seams with 100% literal code moves (A4)
